// apps/backend/src/modules/cms/registry/generated/contract.gen.ts
export type ContractFieldKind =
  | "text" | "html" | "img" | "url" | "color" | "date"
  | "id" | "num" | "bool" | "choice" | "obj" | "list"

export interface ContractField {
  name: string
  kind: ContractFieldKind
  values?: string[]
  fields?: ContractField[]
}